PeoplesGuide.csv
Description:

The People’s Guide is a practical not only is a self-advocacy guide to these self-advocacy information guide and programs, but also gives you advice directory on how to get food, income, on what you can do if your are treated jobs and training, housing, health and unfairly or do not receive what you are dental care, legal advice, and other entitled to by law. The guide is dedicated important help from local, state and to helping all people overcome barriers federal programs and community when they are trying to get help with services in Sacramento County

This is a CSV file of the People's Guide that is in CSV format broken down by headings.